feat: create `get_intermediates_count` on images service
Previously this (ab)used the `get_many` method, and there was a typing issue related to making `limit` and `offset` optional.

Extracted to its own method, fixing the type issue and allowing my IDE to be happy. Also more efficient because we get to use SQLite's optimized `COUNT(*)`.

UI updated.